[feat] add getting lowest price of product by category api

+++ src/main/kotlin/org/tyufjvbn2/productmanage/repositories/BrandRepository.kt
@@ -3,4 +3,4 @@ package org.tyufjvbn2.productmanage.repositories
 import org.springframework.data.jpa.repository.JpaRepository
 import org.tyufjvbn2.productmanage.models.Brand
 
-interface BrandRepository : JpaRepository<Brand, Long> {}
+interface BrandRepository : JpaRepository<Brand, Int> {}
